% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_sv_columns}
\alias{add_sv_columns}
\title{Add Structural Variant Columns}
\usage{
add_sv_columns(metadata, complex = NULL)
}
\arguments{
\item{metadata}{A data.table containing metadata}

\item{complex}{Path to complex events RDS file}
}
\value{
Updated metadata with columns showing the counts of each SV type
}
\description{
Adds a column for each of the structural variant categories from JaBbA complex graph
}
