% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_tmb}
\alias{add_tmb}
\title{Add Tumor Mutation Burden}
\usage{
add_tmb(
  metadata,
  somatic_snvs = NULL,
  jabba_gg = NULL,
  genome = "hg19",
  seqnames_genome_width_or_genome_length = c(1:22, "X", "Y")
)
}
\arguments{
\item{metadata}{A data.table containing metadata}

\item{somatic_snvs}{Path to somatic SNV VCF file}

\item{jabba_gg}{Path to JaBbA graph RDS file}

\item{genome}{The genome reference used}

\item{seqnames_genome_width_or_genome_length}{Sequence names for genome width calculation}
}
\value{
Updated metadata with TMB value added
}
\description{
Calculates and adds tumor mutation burden (TMB) to the metadata based on SNV count and genome length.
}
