% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_treatment_metadata}
\alias{add_treatment_metadata}
\title{Add treatment metadata}
\usage{
add_treatment_metadata(
  metadata,
  input_age_at_biopsy,
  input_treatment_lines,
  input_n_treatment_lines,
  input_best_treatment,
  input_best_treatment_response,
  input_best_treatment_mechanism,
  input_best_treatment_PFS_duration
)
}
\arguments{
\item{metadata}{A data.table containing metadata.}

\item{input_age_at_biopsy}{Age at the biopsy.}

\item{input_treatment_lines}{Different treatment lines received.}

\item{input_n_treatment_lines}{Number of treatment lines received.}

\item{input_best_treatment}{Name of the treatment line with the best response.}

\item{input_best_treatment_response}{Best response obtained among the treatment lines.}

\item{input_best_treatment_mechanism}{Mechanism of the treatment line with the best response.}

\item{input_best_treatment_PFS_duration}{Progression-free survival of the treatment line with the best response.}
}
\value{
Updated metadata with treatment information added.
}
\description{
Adds treatment metadata information such as age at biopsy, treatment lines, number of different treatment lines, and information about the treatment line with the best response (name, response, mechanism, PFS duration).
}
