% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{collect_oncokb_intragenic}
\alias{collect_oncokb_intragenic}
\title{Collect OncoKB intragenic}
\usage{
collect_oncokb_intragenic(
  oncokb_intragenic_deletions = NULL,
  oncokb_intragenic_duplications = NULL,
  pge,
  cytoband,
  verbose = TRUE
)
}
\arguments{
\item{oncokb_intragenic_deletions}{Path to the oncokb intragenic deletions file.}

\item{oncokb_intragenic_duplications}{Path to the oncokb intragenic duplications file.}

\item{verbose}{Logical flag to indicate if messages should be printed.}
}
\value{
A data.table containing processed OncoKB Fusion information.
}
\description{
Collect oncokb intragenic fusions, are parsed separately from the main oncokb fusions file, and include intragenic deletions and duplications.
}
\author{
Kevin Hadi
}
