% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-package.R
\docType{data}
\name{echtvar_vcf_fields}
\alias{echtvar_vcf_fields}
\title{Echtvar annotated vcf fields}
\format{
An object of class \code{character} of length 40.
}
\usage{
echtvar_vcf_fields
}
\description{
Fields from echtvar annotated vcf to keep
}
\keyword{datasets}
