% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{fill_out_normal}
\alias{fill_out_normal}
\title{Fill Out Normal}
\usage{
fill_out_normal(
  cohort,
  normal_cols = c("normal_sample", "normal_bam", "fragcounter_normal",
    "estimate_library_complexity_normal", "alignment_summary_metrics_normal",
    "insert_size_metrics_normal", "normal_wgs_metrics")
)
}
\description{
Create normal with basic information
}
\details{
Need for this arises from gosh output
not always filling out normal paths
if cohort reuses normals across
tumor samples.
}
