% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{lift_filtered_events}
\alias{lift_filtered_events}
\title{lift_filtered_events}
\usage{
lift_filtered_events(
  cohort,
  output_data_dir,
  cores = 1,
  return_table = TRUE,
  additional_columns = Skilift:::default_echtvar_json_fields,
  show_all_tags = FALSE,
  show_all_small_muts_tags = FALSE,
  show_all_cna_tags = FALSE,
  show_all_svs_tags = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\value{
None
}
\description{
Create filtered events for all samples in a cohort
}
