% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{lift_multiplicity}
\alias{lift_multiplicity}
\title{lift_multiplicity}
\usage{
lift_multiplicity(
  cohort,
  output_data_dir,
  is_germline = FALSE,
  node_metadata = c("gene", "feature_type", "annotation", "REF", "ALT", "variant.c",
    "variant.p", "vaf", "transcript_type", "impact", "rank"),
  field = "altered_copies",
  show_only_oncokb = TRUE,
  cores = 1,
  custom_annotation_fields = list(REPORTED = "Reported")
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{is_germline}{Logical indicating if mutations are germline (default: FALSE)}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\value{
None
}
\description{
Create multiplicity JSON files for all samples in a cohort
}
