% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_segmentation_any}
\alias{load_segmentation_any}
\title{Generic segmentation parser}
\usage{
load_segmentation_any(seg_path)
}
\description{
Read any segmentation file or object
}
\author{
GPT 5.4 + Kevin Hadi
}
