% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{merge_oncokb_multiplicity}
\alias{merge_oncokb_multiplicity}
\title{Merge oncokb and multiplicity}
\usage{
merge_oncokb_multiplicity(
  oncokb,
  multiplicity,
  overwrite = FALSE,
  cols.keep = c("annotation", "ref", "alt", "ref_denoised", "alt_denoised", "normal.ref",
    "normal.alt", "variant.g", "major.count", "minor.count", "major_snv_copies",
    "minor_snv_copies", "total_snv_copies", "total_copies", "VAF", "cn",
    "altered_copies"),
  other.cols.keep = NULL
)
}
\description{
Finds overlaps between oncokb maf coordinates and multiplicity
Oncokb is in MAF coordinates which is based on the altered bases,
while multiplicity encodes VCF coordinates based on reference bases.
}
