% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_to_rank}
\alias{order_to_rank}
\title{Order to Rank}
\usage{
order_to_rank(x, ties.method = "average", order_fun = base::order)
}
\description{
Convert order index to rank
}
\details{
Useful for ordering based on multiple factors
}
