% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-helpers.R
\name{sf_duckdb_tables}
\alias{sf_duckdb_tables}
\title{List tables cached in a DuckDB connection}
\usage{
sf_duckdb_tables(duck_con)
}
\arguments{
\item{duck_con}{A DuckDB connection.}
}
\value{
Character vector of table names.
}
\description{
List tables cached in a DuckDB connection
}
