% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\docType{class}
\name{SnowflakeResult-class}
\alias{SnowflakeResult-class}
\alias{show,SnowflakeResult-method}
\alias{dbIsValid,SnowflakeResult-method}
\alias{dbClearResult,SnowflakeResult-method}
\alias{dbFetch,SnowflakeResult-method}
\alias{dbHasCompleted,SnowflakeResult-method}
\alias{dbFetchArrow,SnowflakeResult-method}
\alias{dbFetchArrowChunk,SnowflakeResult-method}
\alias{dbGetRowCount,SnowflakeResult-method}
\alias{dbGetRowsAffected,SnowflakeResult-method}
\alias{dbGetStatement,SnowflakeResult-method}
\alias{dbColumnInfo,SnowflakeResult-method}
\alias{dbGetInfo,SnowflakeResult-method}
\alias{dbBind,SnowflakeResult-method}
\title{SnowflakeResult}
\usage{
\S4method{show}{SnowflakeResult}(object)

\S4method{dbIsValid}{SnowflakeResult}(dbObj, ...)

\S4method{dbClearResult}{SnowflakeResult}(res, ...)

\S4method{dbFetch}{SnowflakeResult}(res, n = -1, ...)

\S4method{dbHasCompleted}{SnowflakeResult}(res, ...)

\S4method{dbFetchArrow}{SnowflakeResult}(res, ...)

\S4method{dbFetchArrowChunk}{SnowflakeResult}(res, ...)

\S4method{dbGetRowCount}{SnowflakeResult}(res, ...)

\S4method{dbGetRowsAffected}{SnowflakeResult}(res, ...)

\S4method{dbGetStatement}{SnowflakeResult}(res, ...)

\S4method{dbColumnInfo}{SnowflakeResult}(res, ...)

\S4method{dbGetInfo}{SnowflakeResult}(dbObj, ...)

\S4method{dbBind}{SnowflakeResult}(res, params, ...)
}
\arguments{
\item{object}{A \linkS4class{SnowflakeResult} object (for \code{show}).}

\item{dbObj}{A \linkS4class{SnowflakeResult} object.}

\item{...}{Additional arguments (currently ignored).}

\item{res}{A \linkS4class{SnowflakeResult} object.}

\item{n}{Number of rows to fetch. -1 (default) fetches all.}

\item{params}{A list of parameter values for parameterized queries.}
}
\description{
An S4 class representing the result of a query against Snowflake.
Created by \code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}} or \code{\link[DBI:dbSendStatement]{DBI::dbSendStatement()}}.
}
\section{Slots}{

\describe{
\item{\code{connection}}{The parent SnowflakeConnection.}

\item{\code{statement}}{The SQL statement string.}

\item{\code{.resp_body}}{The raw JSON response from the first partition.}

\item{\code{.meta}}{Parsed metadata (columns, partitions, handle).}

\item{\code{.state}}{Environment holding mutable state (fetched, partition, valid, etc.).}
}}

