% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_wgd}
\alias{add_wgd}
\title{Add Whole Genome Doubling}
\usage{
add_wgd(
  metadata,
  allelic_jabba_gg = NULL
)
}
\arguments{
\item{metadata}{A data.table containing metadata}

\item{allelic_jabba_gg}{Path to allelic JaBbA graph RDS file}

\item{input_wgd}{Precomputed WGD value}
}
\value{
Updated metadata with WGD value added
}
\description{
Calculates and adds Whole Genome Doubling (WGD) to the metadata.
}
