% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{create_metadata}
\alias{create_metadata}
\title{Create Metadata for a Sample}
\usage{
create_metadata(
  pair,
  tumor_type = NULL,
  tumor_details = NULL,
  disease = NULL,
  primary_site = NULL,
  inferred_sex = NULL,
  purple_pp_bestFit = NULL,
  purity = NULL,
  ploidy = NULL,
  jabba_gg = NULL,
  events = NULL,
  somatic_snvs = NULL,
  germline_snvs = NULL,
  tumor_coverage = NULL,
  foreground_col_name = "foreground",
  estimate_library_complexity = NULL,
  alignment_summary_metrics = NULL,
  insert_size_metrics = NULL,
  tumor_wgs_metrics = NULL,
  normal_wgs_metrics = NULL,
  het_pileups = NULL,
  decomposed_indel_signatures = NULL,
  activities_indel_signatures = NULL,
  matrix_indel_signatures = NULL,
  deconstructsigs_sbs_signatures = NULL,
  decomposed_sbs_signatures = NULL,
  activities_sbs_signatures = NULL,
  matrix_sbs_signatures = NULL,
  hrdetect = NULL,
  onenesstwoness = NULL,
  msisensorpro = NULL,
  genome = "hg19",
  seqnames_loh = c(1:22),
  seqnames_autosomes = c(1:22),
  seqnames_genome_width_or_genome_length = c(1:22, "X", "Y"),
  denoised_coverage_field = "foreground",
  is_visible = TRUE,
  summary = NULL,
  conpair_contamination = NULL,
  conpair_concordance = NULL,
  cohort_type = NULL,
  qc_flags_config = NULL,
  added_field_values = NULL
)
}
\arguments{
\item{pair}{The sample pair identifier.}

\item{tumor_type}{The type of tumor.}

\item{tumor_details}{Details about the tumor.}

\item{disease}{The disease associated with the sample.}

\item{primary_site}{The primary site of the tumor.}

\item{inferred_sex}{The inferred sex of the sample.}

\item{purity}{Purity value (numeric). If provided directly, used instead of deriving from purple_pp_bestFit or jabba_gg.}

\item{ploidy}{Ploidy value (numeric). If provided directly, used instead of deriving from purple_pp_bestFit or jabba_gg.}

\item{jabba_gg}{A file path to the jabba_gg data.}

\item{events}{Structural variant events.}

\item{somatic_snvs}{Somatic single nucleotide variants.}

\item{germline_snvs}{Germline single nucleotide variants.}

\item{tumor_coverage}{Coverage data for the tumor.}

\item{estimate_library_complexity}{Path to the estimate_library_complexity_metrics file.}

\item{alignment_summary_metrics}{Path to the alignment_summary_metrics file.}

\item{insert_size_metrics}{Path to the insert_size_metrics file.}

\item{het_pileups}{Heterozygous pileups data.}

\item{matrix_indel_signatures}{Matrix of indel signatures.}

\item{activities_sbs_signatures}{Activities of SBS signatures.}

\item{matrix_sbs_signatures}{Matrix of SBS signatures.}

\item{hrdetect}{HRDetect scores.}

\item{onenesstwoness}{Oneness and twoness scores.}

\item{msisensorpro}{MSIsensor profile file.}

\item{genome}{The genome reference used.}

\item{seqnames_loh}{Sequence names for loss of heterozygosity.}

\item{seqnames_genome_width_or_genome_length}{Sequence names and genome width in list or genome length as a numeric}

\item{wgs_metrics}{Path to the wgs_metrics file.}

\item{signatures_pair_name}{The name of the signature pair.}
}
\value{
A data.table containing the metadata for a single sample pair.
}
\description{
Creates a comprehensive metadata object for a single sample pair by aggregating various data inputs.
}
