% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\docType{data}
\name{default_col_mapping}
\alias{default_col_mapping}
\title{Default column mappings}
\format{
An object of class \code{list} of length 91.
}
\usage{
default_col_mapping
}
\description{
Defines naming convention cohort inputs table
}
\details{
This variable contains the possible regex patterns
from an input data.table with columns corresponding to processed data.
You can output the default template to modify it using
base::dput(Skilift::default_col_mapping)
}
\keyword{datasets}
